+++ kernel/rlite-kernel.h
@@ -215,7 +215,7 @@ static inline int
 rl_buf_custom_pop(struct rl_buf *rb, size_t len)
 {
     if (unlikely(rb->len < len)) {
-        RPD(1, "No enough data to pop %d bytes\n", (int)len);
+        RPD(1, "No enough data to pop %zu bytes\n", len);
         return -1;
     }
 
@@ -229,7 +229,7 @@ static inline int
 rl_buf_custom_push(struct rl_buf *rb, size_t len)
 {
     if (unlikely((uint8_t *)(rb->pci) - len < &rb->raw->buf[0])) {
-        RPD(1, "No space to push %d bytes\n", (int)len);
+        RPD(1, "No space to push %zu bytes\n", len);
         return -1;
     }
 
@@ -302,7 +302,7 @@ static inline int
 rl_buf_custom_pop(struct rl_buf *rb, size_t len)
 {
     if (unlikely(rb->len < len)) {
-        RPD(1, "No enough data to pop %d bytes\n", (int)len);
+        RPD(1, "No enough data to pop %zu bytes\n", len);
         return -1;
     }
 
@@ -315,7 +315,7 @@ static inline int
 rl_buf_custom_push(struct rl_buf *rb, size_t len)
 {
     if (unlikely(skb_headroom(rb) < len)) {
-        RPD(1, "No space to push %d bytes\n", (int)len);
+        RPD(1, "No space to push %zu bytes\n", len);
         return -1;
     }
